+++ plutarch-ledger-api/src/Plutarch/LedgerApi/V1/MintValue.hs
@@ -26,6 +26,9 @@ import PlutusTx.Prelude qualified as PlutusTx
 {- | Represents sorted, well-formed Values with a mandatory /zero/ Ada entry,
 while all other token quantities must be non-zero.
 
+Duplicate currency symbols or duplicate token names within the same
+token map are not allowed (since wip).
+
 @since 3.5.0
 -}
 newtype PMintValue (s :: S) = PMintValue (Term s PSortedValue)
@@ -99,11 +102,17 @@ instance PValidateData PMintValue where
     plet (pfromData $ pparseData @PSortedValue opq) $ \value ->
       pif
         ( (pnot #$ phasZeroAdaEntry # value)
-            -- TODO: rewrite this check without adding fake ADA
-            #|| (phasZeroTokenQuantities #$ pforgetSorted $ value <> (psingletonSortedValue # padaSymbol # padaToken # 1))
+            #|| (phasZeroTokenQuantities #$ pforgetSorted $ dropAdaEntry # value)
         )
         perror
         x
+    where
+      dropAdaEntry :: forall (s :: S). Term s (PSortedValue :--> PSortedValue)
+      dropAdaEntry =
+        plam $ \value ->
+          pmatch (pto $ pto $ pto value) $ \case
+            PNil -> value
+            PCons _ xs -> punsafeDowncast $ punsafeDowncast $ punsafeDowncast xs
 
 {- | Construct an empty 'PMintValue' with a zero Ada entry.